refactor: extract PieChartDrawer from PieChartView
